refactor: navigateToPurchasePage 함수 따로 파일 분리
