+++ components/shippingAddress.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const shippingAddress = () => {
+  return <div>shippingAddress</div>;
+};
+
+export default shippingAddress;
