+++ lib/utils/navigateToPurchasePage.ts
@@ -1,5 +1,10 @@
 import { useRouter } from 'next/navigation';
 
+/**
+ * 구매 페이지로 이동하는 함수
+ * @param param raffleId
+ * @returns
+ */
 export default function navigateToPurchasePage({ raffleId }: { raffleId: string }) {
   const router = useRouter();
   return () => router.push(`/purchase/${raffleId}`);
